+++ hilti/toolchain/include/ast/node-tag.h
@@ -489,8 +489,9 @@ constexpr Tag Unequal = 3406;
 constexpr Tag Split = 3407;
 constexpr Tag Split1 = 3408;
 constexpr Tag StartsWith = 3409;
-constexpr Tag LowerCase = 3410;
-constexpr Tag UpperCase = 3411;
+constexpr Tag EndsWith = 3410;
+constexpr Tag LowerCase = 3411;
+constexpr Tag UpperCase = 3412;
 } // namespace string
 
 namespace strong_reference {