+++ hilti/toolchain/include/ast/visitor-dispatcher.h
@@ -374,6 +374,7 @@ public:
     virtual void operator()(hilti::operator_::string::Split* n) {}
     virtual void operator()(hilti::operator_::string::Split1* n) {}
     virtual void operator()(hilti::operator_::string::StartsWith* n) {}
+    virtual void operator()(hilti::operator_::string::EndsWith* n) {}
     virtual void operator()(hilti::operator_::string::LowerCase* n) {}
     virtual void operator()(hilti::operator_::string::UpperCase* n) {}
     virtual void operator()(hilti::operator_::struct_::Unset* n) {}